feat(video): Phase 2B adaptive player UI, docs, and app updates
- Portrait/landscape controls, LTR seek, speed/aspect/subtitle chrome
- Locked/loading/error overlays; notifier playback speed sync on open
- Docs: mapper, VIDEO_AGENT, ROADMAP; theme/shell/router and cross-feature fixes

+++ lib/features/video_player/domain/entities/gesture_engine.dart
@@ -9,39 +9,39 @@ import 'gesture_result.dart'; // For GestureResult
 enum GestureType { none, seek, volume, brightness }
 
 class GestureEngine {
-  GestureType _type     = GestureType.none;
-  bool        _isLocked = false;
-  Duration    _preview  = Duration.zero;
-  double      _vol      = 0.0;
-  double      _bright   = 0.0;
+  GestureType _type = GestureType.none;
+  bool _isLocked = false;
+  Duration _preview = Duration.zero;
+  double _vol = 0.0;
+  double _bright = 0.0;
 
   // ── Configurable Parameters ────────────────────────────────
-  final double threshold;       // default: 8.0 px
-  final double fastThreshold;   // default: 10.0 px/frame
-  final double seekSlowMs;      // default: 400.0 ms/px
-  final double seekFastMs;      // default: 1200.0 ms/px
+  final double threshold; // default: 8.0 px
+  final double fastThreshold; // default: 10.0 px/frame
+  final double seekSlowMs; // default: 400.0 ms/px
+  final double seekFastMs; // default: 1200.0 ms/px
   final double vertSensitivity; // default: 0.01 (1% per pixel)
 
   GestureEngine({
-    this.threshold      = 8.0,
-    this.fastThreshold  = 10.0,
-    this.seekSlowMs     = 400.0,
-    this.seekFastMs     = 1200.0,
-    this.vertSensitivity= 0.01,
+    this.threshold = 8.0,
+    this.fastThreshold = 10.0,
+    this.seekSlowMs = 400.0,
+    this.seekFastMs = 1200.0,
+    this.vertSensitivity = 0.01,
   });
 
   void onStart({
-    required double   dx,
-    required double   screenWidth,
+    required double dx,
+    required double screenWidth,
     required Duration currentPosition,
-    required double   volume,
-    required double   brightness,
+    required double volume,
+    required double brightness,
   }) {
-    _preview  = currentPosition;
-    _vol      = volume;
-    _bright   = brightness;
+    _preview = currentPosition;
+    _vol = volume;
+    _bright = brightness;
     // Tentative vertical type based on screen half
-    _type     = dx < screenWidth / 2 ? GestureType.brightness : GestureType.volume;
+    _type = dx < screenWidth / 2 ? GestureType.brightness : GestureType.volume;
     _isLocked = false;
   }
 
@@ -62,7 +62,8 @@ class GestureEngine {
     // Phase 2: execute on locked type
     switch (_type) {
       case GestureType.seek:
-        final speed = d.delta.dx.abs() > fastThreshold ? seekFastMs : seekSlowMs;
+        final speed =
+            d.delta.dx.abs() > fastThreshold ? seekFastMs : seekSlowMs;
         _preview += Duration(milliseconds: (d.delta.dx * speed).toInt());
         if (_preview <= Duration.zero) {
           _preview = Duration.zero;
@@ -74,11 +75,12 @@ class GestureEngine {
         return GestureResult.seek(_preview);
 
       case GestureType.volume:
-        _vol = (_vol - d.delta.dy * vertSensitivity).clamp(0.0, 1.0);
+        _vol = (_vol - d.delta.dy * vertSensitivity).clamp(0.0, 1.0).toDouble();
         return GestureResult.volume(_vol);
 
       case GestureType.brightness:
-        _bright = (_bright - d.delta.dy * vertSensitivity).clamp(0.0, 1.0);
+        _bright =
+            (_bright - d.delta.dy * vertSensitivity).clamp(0.0, 1.0).toDouble();
         return GestureResult.brightness(_bright);
 
       default:
@@ -87,7 +89,7 @@ class GestureEngine {
   }
 
   void reset() {
-    _type     = GestureType.none;
+    _type = GestureType.none;
     _isLocked = false;
   }
 }
